Refactor App component structure to integrate ThemeProvider and SearchProvider, and update navigation path in LoginForm
