+++ frontend/src/main.tsx
@@ -1,15 +1,15 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import AppProvider from "./components/Layout/AppProvider";
 import { AuthProvider } from "./context/AuthContext";
 import AppRoutes from "./routes";
+import App from "./App";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <AuthProvider>
-      <AppProvider>
+      <App>
         <AppRoutes />
-      </AppProvider>
+      </App>
     </AuthProvider>
   </React.StrictMode>
 );
